+++ apps/app/src/test/scala/com/daml/network/integration/tests/GlobalDomainUpgradeTimeBasedIntegrationTest.scala
@@ -211,7 +211,6 @@ class GlobalDomainUpgradeTimeBasedIntegrationTest
                 "alice-participant-id",
                 dummyRound,
                 previousGlobalId.toProtoPrimitive,
-                sv1Backend.config.domainMigrationId,
               )
             )
           ),