+++ cluster/pulumi/common/src/domainFees.ts
@@ -4,8 +4,12 @@ export type ValidatorTopupConfig = {
 };
 
 // TODO(#6032): determine the best defaults here
-export const domainFeesConfig = {
-  // Validator
+export const svValidatorTopupConfig: ValidatorTopupConfig = {
+  targetThroughput: 0,
+  minTopupInterval: '1m',
+};
+
+export const nonSvValidatorTopupConfig: ValidatorTopupConfig = {
   targetThroughput: 33330, // 10x the base rate to reduce the chance of validators locking up (see #6313)
   minTopupInterval: '1m',
 };